+++ libnail/CHANGELOG.md
@@ -18,10 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - added fields `profile_name`, `target_name` to struct `AlignmentBuilder`
-- added struct `AlignmentStats` 
+- added struct `AlignmentStats`
+- added variants `Length`, `Pid`, `MatchCount`, `MismatchCount`, `GapOpenCount`, `GapCount`, to enum `Field`
+- added field `stats: AlignmentStats` to struct `Alignment` 
 
 ### Changed
-- struct `Alignment` now has field `stats: AlignmentStats`
 - struct `DisplayStrings` now derives `PartialEq`
 - function `map_posterior_probability_to_bin_byte()` renamed to `map_posterior_probability_to_bin_byte_utf8()`